+++ R/ImageArray.R
@@ -14,7 +14,6 @@
 #'  }
 #' }
 #' @param drop ignored
-#' @param angle value between 0 and 360 for degrees to rotate
 #' @param brightness the brightness of the new image in percentage, e.g. 120
 #' @param perm perm
 #' @param index a named or unnamed list of indices for cropping/subsetting the
@@ -30,14 +29,8 @@
 #' @aliases
 #' [[,ImageArray,numeric-method
 #' [[<-,ImageArray,numeric-method
-#' rotate
-#' rotate,ImageArray-method
 #' crop
 #' crop,ImageArray-method
-#' flip
-#' flip,ImageArray-method
-#' flop
-#' flop,ImageArray-method
 #' negate
 #' negate,ImageArray-method
 #' modulate
@@ -52,6 +45,8 @@
 #' as.raster,ImageArray-method
 #' path
 #' path,ImageArray-method
+#' extent
+#' extent,ImageArray-method
 #'
 #' @examples
 #' # get image
@@ -68,6 +63,9 @@
 #' # dimensions and length
 #' dim(imgarray)
 #' length(imgarray)
+#' 
+#' # extent
+#' extent(imgarray)
 #'
 #' # manipulate images
 #' imgarray <- crop(imgarray, ind = list(100:200, 100:200))
@@ -99,12 +97,6 @@ setMethod(
   f = "[",
   signature = c("ImageArray"),
   function(x, i, j, ..., drop = FALSE) {
-    if (missing(x)) {
-      stop("'x' is missing")
-    }
-    if (!.isTRUEorFALSE(drop)) {
-      stop("'drop' must be TRUE or FALSE")
-    }
     Nindex <- S4Arrays:::extract_Nindex_from_syscall(sys.call(), parent.frame())
     crop(x, index = Nindex)
   }
@@ -226,7 +218,8 @@ createBFArray <- function(
   image_list <- lapply(resolution, function(res) {
     BFArray(image, series = series, resolution = res)
   })
-  ImageArray(meta = list(axes = tolower(names(image_list[[1]]@seed@shape))), levels = image_list)
+  ImageArray(meta = list(axes = tolower(names(image_list[[1]]@seed@shape))), 
+             levels = image_list)
 }
 
 #' createMagickArray
@@ -580,17 +573,10 @@ writeImageArray <- function(
     if(format == "hdf5") format <- "h5"
     if(fileext != format && format != "in-memory") {
       warning(
-        sprintf(
-          paste(
-            "The file extension of the output path%s does", 
-            "not match the specified format (%s),", 
-            "The object will be written as (%s).", 
-            sep = " "
-          ),
-          if (fileext == "") "" else paste0(" (", fileext, ")"),
-          format,
-          format
-        )
+        "The file extension of the output path", 
+        if (fileext == "") "" else sprintf(" '%s'", fileext),
+        " does not match the specified format (", format, "). ", 
+        "The object will be saved as ", format, " format. "
       )
     }
   }