+++ tests/testthat/test-manipulation.R
@@ -44,15 +44,6 @@ test_that("manipulate h5 ImageArray", {
   expect_equal(unique(as.vector(tmp)), 255)
   expect_equal(type(mat_list_negated[[1]]), "integer")
 
-  # rotate
-  mat_list_rotated <- rotate(mat_list, angle = 90)
-  expect_equal(dim(mat_list_rotated), c(3, 2000, 5000))
-  mat_list_rotated <- rotate(mat_list, angle = 180)
-  expect_equal(dim(mat_list_rotated), c(3, 5000, 2000))
-  mat_list_rotated <- rotate(mat_list, angle = 270)
-  expect_equal(dim(mat_list_rotated), c(3, 2000, 5000))
-  expect_error(mat_list_rotated <- rotate(mat_list, angle = 20))
-
   # flip flop
   mat_list_flipflop <- flip(mat_list)
   expect_equal(
@@ -118,15 +109,6 @@ test_that("manipulate zarr ImageArray", {
   mat_list_negated <- negate(mat_list)
   tmp <- realize(mat_list[[1]]) + realize(mat_list_negated[[1]])
   expect_equal(unique(as.vector(tmp)), 255)
-
-  # rotate
-  mat_list_rotated <- rotate(mat_list, angle = 90)
-  expect_equal(dim(mat_list_rotated), c(3, 2000, 5000))
-  mat_list_rotated <- rotate(mat_list, angle = 180)
-  expect_equal(dim(mat_list_rotated), c(3, 5000, 2000))
-  mat_list_rotated <- rotate(mat_list, angle = 270)
-  expect_equal(dim(mat_list_rotated), c(3, 2000, 5000))
-  expect_error(mat_list_rotated <- rotate(mat_list, angle = 20))
 
   # flip flop
   mat_list_flipflop <- flip(mat_list)