+++ NEWS.md
@@ -1,6 +1,15 @@
+# ImageArray 1.1.3
+
+## New features
+
+* Novel delayed (or lazy) transformations for image pyramids are introduced. 
+  Available transformations are: Translation, scaling, rotation, and affine 
+  (#45).
+* New `extent` function returns the physical space that the image occupies.
+  
 # ImageArray 1.1.2
 
-## NEW FEATURES
+## New features
 
 * Images with any combination of XYZCT dimensions are can now be read.
   See OME webpage for more information: